supernova/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
